requirements.txt
pandas
bson
openai
pymongo
matplotlib.pyplot